backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.19
authlib==1.3.2
httpx==0.28.1
openai==1.57.0
stripe==11.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
celery==5.4.0
redis==5.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
